requirements.txt
Flask==2.0.2
Turbo-Flask==0.6.1
pytube==11.0.1
opencv-python==4.5.3.56
pytesseract==0.3.8
pandas==1.3.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
loguru==0.5.3
img2pdf==0.4.1
imutils==0.5.4
numpy==1.21.2
fire==0.4.0
